Move to a safe-globals package
