+++ packages/safe-globals/index.js
